Update wifi button step number
Update the number of steps that the current position is changed by when a button is pressed on the wifi interface.
With the current settings the press of a button changes the position along the specific axis by full range / 20. Meaning that to go from on end of the axis range to the other, a button has to be pressed 20 times.